vkconfig: refactor doc generation into dedicated code file
